+++ ncvue/__init__.py
@@ -18,7 +18,9 @@ from .version import __version__, __author__
 # helper functions
 # general helper function
 from .ncvutils   import DIMMETHODS, SEPCHAR
-from .ncvutils   import add_cyclic_point, clone_ncvmain, get_slice
+from .ncvutils   import add_cyclic_point, clone_ncvmain
+from .ncvutils   import format_coord_contour, format_coord_map
+from .ncvutils   import format_coord_scatter, get_slice
 from .ncvutils   import list_intersection, set_axis_label, set_miss
 from .ncvutils   import spinbox_values, vardim2var, zip_dim_name_length